Turing machine: remove useless guards
The search of an operator occurs by reading the band from left to right. Rewinding the band occurs to search for its operands. When done, the next operator is searched from the handled operator.

The initial string is always completely processed when the last operator is handled.

A well-formed string has no operand after the last operator.